Removing unnecessary ActiveSupport::Concern dependency
